tests: split by groups
